[netstack3] Remove try_unit
This macro is a shorthand that saves a few lines at the expense of
readability; removing it also revealed a number of inconsistent usages
with or without Option::ok_or_else.

This macro also bakes-in Result::ok, which is contrary to networking
rust patterns.

Remove `impl From<Ipv6SourceAddr> for Option<UnicastAddr<Ipv6Addr>>` and
`Ipv6SourceAddr::to_option` as the former is unused and the latter has a
single remaining caller which is trivially removed.

Change-Id: Iaa638a8bddd1172a4862ea097861e2b4e494497b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620784
Reviewed-by: Joshua Liebow-Feeser <joshlf@google.com>
Fuchsia-Auto-Submit: Tamir Duberstein <tamird@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>