[assembly] Allow the blob.manifest to be located outside CWD
The blob.manifest contains paths to blobs relative to the blob.manifest
itself. Previously, the assembly code required that the blob.manifest
was located underneath the CWD hierarchy of files. In order to support
out-of-tree workflows, this does not work, therefore we are making it
more flexible.

Bug: 85370
Change-Id: Ie5c5a0a6f620bc423c5bbdfdf3d1334a0b04739d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621901
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>