[netstack] Shorten autogenerated test case name
ResultDB sets a limit of 512 characters for a test case name. When
no base name is given to the test_case macro, it autogenerates a name
using the remaining contents. In this case, the autogenerated name is
over 500 characters and is close to exceeding the limit of what
ResultDB will accept.
This change specifies a base name to greatly shorten the autogenerated
test case name.

Testing: Manually run fx test netstack-inspect-integration-test with
and without this change and verify the same number of tests are run.

Change-Id: I207cca497c0f2bf426e3421a4695330b6c0c770e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621902
Fuchsia-Auto-Submit: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>