[kernel][vm] Assert protect unmap result
Asserts that the result of protect and unmap is ZX_OK instead of
returning the status. The AspaceUnmapVmoRangeLocked and
AspaceRemoveWriteVmoRangeLocked methods do not actually have their
return values checked, and are called from methods that return void
anyway. Explicitly asserting ensures that in the unlikely event these
methods fail we do not continue with an incorrect state.

Change-Id: I6f8315e269d1d9687de888f2a4698ce17be0980b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/615947
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>