[kernel][vm] Make over unmapping explicit
Extends ArchVmAspace::Unmap to take a parameter that controls whether
or not unmap can operate on a larger range than requested. A larger
range might be unmapped if a large page needs to be split, and this
can fail if there's no memory for a page table. For the kernel aspace
over unmapping is not tolerable, since the kernel cannot fault the
the pages back in later, unlike a user address space.

Change-Id: Id6ecfd5a46dcd36567708469a7e604bad21c88fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/615948
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>