[realm_builder][rust] implement client side of new API
Introduces a new implementation for the RealmBuilder struct that is
based on the new fuchsia.component.test/Realm and /Builder protocols,
instead of the old fuchsia.component.test/RealmBuilder protocol.

This new implementation is in a new module named `soft_migration`. After
this lands clients will migrate to the new definitions under
`soft_migration`, and once this is complete the definitions for the new
implementation will be moved to the top-level of this crate (and
reexported under the old soft_migration module until clients migrate
back to the top-level definitions).

Change-Id: I821e8c0afc5033c1de815162af300e66c545c177
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613645
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Derek Gonyeo <dgonyeo@google.com>