[fxfs][fsck] Check extents
Ensures that allocation and extent records are aligned and well-formed
(start < end).

Also fixed a few places where the production code panics if it
encounters a bad range, in favour of errors (which still sometimes
panic, but at least we have the option to do otherwise).

Bug: 87381
Change-Id: Iac6f30e017b8008a3ff5b3425561ab01c7ea7a6f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620888
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>