[fxfs][fsck] Make fsck continue on some errors
Adds support for a third variant of errors which fsck should keep going
on, but which are still failures by the time fsck is finished.  This
leaves three categories:
- Warning: Log the errors, but don't fail fsck.
- Error: Fail fsck, but keep it running to collect more errors.
- Fatal: Immediately fail fsck.

Move many of the existing errors to behave as Errors, keeping only
really serious ones as Fatal.

Also adds an extra check that Volume records are only found in the root
store (which motivated these other tweaks).

Bug: 87381
Change-Id: Ic8d5d62914ede6b758655a3fe77fe66fab33fc55
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621761
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>