[camera] Add a name field to VmoPoolWrapper
Add a name field to VmoPoolWrapper, which is optionally set in the
Init() method. If set, the name is displayed in syslog lines for the
VmoPoolWrapper, making it easier to differentiate between different
instances.

Also update GDC and GE2D tasks to set their VmoPoolWrapper's name.

Bug: b/210170011
Change-Id: I547a1346226581460fc1f9e204eee77218ca3275
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621263
Reviewed-by: Matt Sandy <msandy@google.com>
Commit-Queue: Marcus Jang <marcusjang@google.com>