[virtualization][linux_runner] Switch to bullseye image
Bullseye is now the latest stable release of debian.

This also avoids using the debian version in the
container name as that requires it to be updated
each time we switch to a new version of debian. Also
going to be confusing when upgrading containers.

Fixes: 90525
Change-Id: I300a78d38a46c99756087236b059625bee54165e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621381
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: David Reveman <reveman@google.com>