[botanist] Fix race between runAgainstTarget() and target.Wait().
When runAgainstTarget() completes, it shuts down the target. However,
the target calls Wait() in another goroutine to see if it gets any other
errors so that it can exit early. If the target gets shut down properly,
but we receive the target.Wait() error before getting the
runAgainstTarget() error, then we could falsely fail the task.
Cancelling the errgroup context once runAgainstTarget() is finished and
checking for the context error in the other goroutines should prioritize
the error we get from runAgainstTarget() in the case there is a race.

Bug: 90679
Change-Id: I9d892149b226792a96e4e7feb2ba85cff69f2efe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/623201
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>