[sshutil] Fix noisy logs
"error closing keepalive session" and "closing ssh session: EOF" would
previously be printed in nearly every task, whether it passed or failed,
because session.Close() returns an error if the connection has already
intentionally been closed by testrunner or botanist. So if Close() fails
because the session was already closed, it means we're shutting down
normally so we should ignore the error, causing it to not be logged.

Bug: 90137
Change-Id: I8b398639b723ec69e479095878e0f878a0b6e242
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/623222
Reviewed-by: Ina Huh <ihuh@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>