[audio_core][test] Fix flaky audio clock test case
Clock tests are written to skip the actual synchronization
checks if any kind of underflow or overflow occur, but also
to automatically fail to bring attention to those underflows.
The cause of these underflows continues to be investigated
and addressed (tracked by fxb/80003), so in the meantime we
only automatically fail on Renderer underflows, based on the
assumption that the test cases submit the entire source
signal before starting playback.

Unfortunately this is not the case for these test cases, so
all underflow checks have been disabled until fxb/80003 is
fixed.

Multiply: audio-clock-pipeline-test (fuchsia)
Fixed: fxb/79916

Change-Id: If0e38955201adf4640c1dbe143dd2b20e21b32c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/623289
Reviewed-by: Dale Sather <dalesat@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>