[pretty] Update unit k -> K
The NIST(National Institute for Standard and Technology) defines:
kibi for kilobyte with KiB as symbol.
https://physics.nist.gov/cuu/Units/binary.html

The kB can be interpreted as Kilobytes or kibi (1000 or 1024).

The library is consistent with powers of 2, which is why it should
output units with an uppercase K. Even the cpp SizeUnit is KiB.

This change also makes the unit consistent with the rest.

Interesting stack overflow discussion regarding the subject, includes
references to standard as well:
https://superuser.com/questions/938234/size-of-files-in-windows-is-it-kb-or-kb

No functional changes.

Test: CQ
Change-Id: Ib59d91fe47e82556340665ffa9d8376e2d69d9a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/623282
Reviewed-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>