[usb][cdc] cdc ether prevents host from suspending
suspend should be handled on only on the usbsusp signal, not the
erlysusp signal as well.

deal with GET_STATUS request on host bootup.

this change allows the host to suspend, but the host does not stay
suspended possibly due to bluetooth wakeup events.

Bug: 66889

Test: tested on device

Change-Id: I2f8b38a1a17813abf761a83285782ede405aa732
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609466
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Ruby Zhuang <rdzhuang@google.com>