[zircon][time] Remove ZX_CLOCK_UTC.
This enumeration was defined mainly for use in the zx_clock_get syscall
that was deleted ~6 months ago in accordance with RFC-0008. We had a
few additional usages, particularly of the derived zx::time_utc type,
but have steadily burnt those down over the last couple of months and
we're ready to remove the constant and with it the last vestige of UTC
as a kernel responsibility!

Bug: 86976
Change-Id: I07ebfa56572cea7cf31e0e222660d23b41b1f75e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614036
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>