[netstack] Remove a panic from address removal
Removing an address used to do lookups to ensure that the interface
and the address exists, and then panic if `stack.RemoveAddress` returns
an error that indicates that the interface no longer exists.

Don't do the lookups, since it's possible for the interface to be
removed between doing the lookup and calling `stack.RemoveAddress`.

Add a regression integration test which removes addresses and the
interface to which the addresses belong at the same time.

Fixed: 90396
Test: fx test --test-filter 'remove_interface_and_address_*' \
              netstack-fidl-integration-test

Change-Id: I6ff572421d0354cf3e0dae81cbb5c6b74fe4a015
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619961
Reviewed-by: Amanda Tait <atait@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Tony Gong <gongt@google.com>