[llcpp] Remove Transport parameter from ClientEnd and ServerEnd
This CL removes the transport parameter from ClientEnd and ServerEnd and
instead retrieves the transport from a using declaration on the protocol
marker. A field on the protocol marker was used for this rather than a new
trait because the transport needs to be accessible in generated wire type
files.

Additionally, this CL:
- Removes dependencies on event and sync client template logic for the
driver transport. This reduces the amount of existing channel-specific
code that needs to change.
- Adds a dependency on the driver_runtime for LLCPP libraries in
fidl_cpp.gni. This is temporary and was necessary to resolve dependency
issues (wire types depend on ClientEnd/ServerEnd which depends on
the driver transport).

Change-Id: I80fbe8b7cc30445e548e44bba05e750b70744a57
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620923
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
API-Review: Ian McKellar <ianloic@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>