[component_manager] Destroy event stream task on component purge
In EventStreamProvider, the task that was serving the event stream over
FIDL was being detached. This led to an issue where if the component
was purged, the serving task would continue running.

This resulted in a handle + memory leak in CM.

The task has now been tied to the component lifecycle by adding it to
EventStreamAttachment.

EventStreamAtttachments for a component are cleaned up on Purge.

Bug: 80731
Test: fx set core.x64 --with //src/sys/component_manager:tests
      fx test
Change-Id: Icf35ffcc1657b10f0befe343bbd696bbb3ca7547
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/623030
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>