[blobfs] add new component runner
Add a new Runner class to blobfs for running in a component context.
Running as a component requires us to parameterize blobfs in a slightly
awkward way. When it's first launched, it serves a new fidl protocol,
fuchsia.fs.startup.Startup, which takes a block device and launch
arguments for launching the filesystem, as well as providing fsck- and
mkfs-like functions. Once the launch function is called, the filesystem
is configured, which represents a second stage of startup procedure.

This also adds a new blobfs-component binary which launches blobfs using
this new runner, and a component-level integration test for it.

Bug: 82896
Test: fx test blobfs-component-runner-test
Change-Id: I557f979fe8805740b449b99c151b1121ba9d9b07
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620886
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
API-Review: Chris Suter <csuter@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>