[debugger] Improve syntax highlighting.
Adds syntax highlight support for lines beginning with # which are
formatted like a comment. This handles C preprocessor and Rust
attributes and improves source code listings somewhat.

Adds tokenizer support for Rust 'lifetime tokens. This is not needed to
support any expressions but comes up regularly in syntax highlighting.
Without this patch, any source line with a lifetime in it is
unhighlighted. With this, lifetimes are highlighted like keywords and
any other tokens on the same like are also supported.

Change-Id: I184abf1366f3599308e8f94eb0600e3eecd2633b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621753
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>