[camera] Add deallocation failure fallback
This change adds a fallback path in camera device which allows changing
the configuration even if waiting for deallocation of previous stream
collections failed. This reduces the risk of failure due to exhaustion
of contiguous memory in cases where the wait failed.

This change also fixes an instance of use-after-move which could trigger
the failure mode above, and adds additional logging during the connection
process to identify failures.

MULTIPLY: `[
  {
    "name": "camera_device_unittest",
    "total_runs": 100
  }
]`
Bug: b/199321382
Test: ran unit tests and product scenarios locally

Change-Id: Icaa335ca9a61c3dac712a6f0da842d83017fbf3d