[bt][usb] Cancel requests in Unbind
Cancel pending USB requests in DdkUnbind(). Waiting for requests to
fail is insufficient, and is likely to block.

Remove pending_requests_completed_ condition as it is no longer
necessary. usb_cancel_all is synchronous.

In DdkUnbind(), spawn a thread for the unbind routine in order to avoid
blocking the main thread.

Bug: 90277
Test: fx test bt-transport-usb-tests
Change-Id: Id71afe642b07fc067f9528a605bca902f5e9041f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/624307
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Ben Lawson <benlawson@google.com>