[intel-i915] Pipe: inject controller at use time instead of in ctor.
Instead of keeping a pointer to |i915::Controller| in Pipe instances
all the time and initialize |Pipe| using Controller, we can only
keep the MMIO space of the controller, and inject controller or
controller methods only when it is needed.

This could help us test Pipe without creating a full-fledged
i915 Controller object, which makes it easier for unit testing.

Bug: 83998
TEST: no-op, refactoring only.

Change-Id: Ib9a63984ef2697b99523bb98cd8a8ec6897005b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/622904
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: Yilong Li <liyl@google.com>