[display] Use configuration stamps in ApplyConfig and Vsync events.
For each display configuration applied to display Controller
service, it will be associated with a unique tag (config stamp).

Display controller clients could call GetLatestAppliedConfigStamp()
to get the config stamp for previous ApplyConfig().

On each hardware Vsync signal, the display controller will send a
OnVsync2() event to its active client. The event contains a config
stamp means that all display layers has fully applied the configuration
with that stamp.

If there are pending images in a configuration, the configuration
stamp in Vsync will be only updated after the pending images are
resolved, or the layer is removed in a subsequent config, or the
layer gets a new config that overrides this configuration.

TEST=Added new test cases in display integration-tests.
Bug: 72588

Change-Id: Ieef3b785eecf3085ee0c08577be2abdb27d97732
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620345
Reviewed-by: Arman Uguray <armansito@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
API-Review: John Bauman <jbauman@google.com>
Commit-Queue: Yilong Li <liyl@google.com>