[component_manager] Add a debug configuration for nested component manager
When a nested component manager is started by RealmBuilder, it is
usually using a configuration + manifest that does not enable debug
mode.

This means that tests cannot subscribe to events before starting the
component topology.

In the transition from OpaqueTest, we will need a config file and
manifest for a component manager that uses RealmBuilder and is in
debug mode.

Bug: 71873
Test: N/A
Change-Id: If0e6d8bb3c2736ddf2ac71cd47067b35bb3a636f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/624223
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>