[display] Introduce SetDisplayPower.
Hardware-specific drivers can implement this as a no-op for now, and
this particular API will evolve as we discover new requirements.

The core display driver will also be updated with a new API, but it is
not yet clear whether to modify fuchsia.hardware.display.Controller or
to introduce a secondary protocol.

Change-Id: I75f2490401a91c03a77593cdc3cd8209a4939ee4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573565
Reviewed-by: Yilong Li <liyl@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
API-Review: John Bauman <jbauman@google.com>
Commit-Queue: Randall Bosetti <rlb@google.com>