[component][fuzzing] Add fuzz-manager
This CL adds an implementation of the fuchsia.fuzzer.Manager protocol.
This service will be available on startup when included in a build. It
accepts requests from remote-control and starts fuzz tests using the
test-manager service.

Additionally, the fuzz-manager bridges the diagnostic streams such as
stdout, stderr and syslog from the test-manager to clients. It buffers
this data to allow clients to disconnect and reconnect as part of
long-running fuzzing sessions.

Bug: 24740
Change-Id: I3c01a7e8c1e836aa34a542c529da31c098b36bd8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/558391
Reviewed-by: Chris Palmer <palmer@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>