[component][fuzzing] Add fuzz-registry
This CL adds an implementation of fuchsia.fuzzer.Registry. The
fuzz-test-runner adds one end of a fuchsia.fuzzer.ControllerProvider
channel to the registry when starting a fuzz test. The fuzz manager
can then forward requests for fuchsia.fuzzer.Controller to this
service.

The implementation of this protocol is in a separate component to avoid
creating a circular dependency between the fuzz-manager, the
test-manager, and the fuzz-test-runner.

Bug: 24740
Change-Id: I3a54c9b2557a4b8cfe8a066801a166396efb11cd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/580183
Reviewed-by: Chris Palmer <palmer@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>