[testrunner] Add NsJail wrapper
Adds a thin wrapper around NsJail that will facilitate the construction
of command line invocations. Also adds a flag to store a passed in
NsJail binary path. This CL does not actually wire up the usage of said
flag or wrapper, it just puts the pieces in place for a future CL to
make use of them. This was done to enable CQ testing of NsJail changes.

Bug: b/211504655
Change-Id: I05e007b0b63d07e0f7247b182a72ce3454184483
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/624683
Reviewed-by: Ina Huh <ihuh@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>