[blobfs] queue incoming requests until blobfs is started
This changes the component runner to queue incoming requests to the root
of the filesystem and the svc directory until the startup protocol is
called. It does this by creating dangling channel pairs which it puts in
it's outgoing directory at /root and /svc, which it only starts serving
once the Configure callback is called by the startup service. The
diagnostics directory isn't expected to adhere to this behavior, and so
it doesn't get this treatment.

This also adds a test which verifies the request queuing behavior to
both the blobfs integration tests and the blobfs-component tests.

Test: fx test blobfs-component-runner-test blobfs-component-test
Bug: 82896
Bug: 39588
Change-Id: I464b9402a1807a2e7aa16d2f429508f4bf52a34c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/625747
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>