[ffx] Add template rendering for ffx emu flags
This CL adds the template file for ffx emu command line arguments, and
the code for parsing and utilizing the template to build the command
line.

We use Handlebars to provide variable substitution and conditional
inclusion of certain flags in the template. Two new handlebars helpers
are defined, one for verifying the value of an enumeration, and the
other to call the abbreviate() function on a DataUnits member.

Bug: fxbug.dev/90481

Change-Id: Ic79095718ea7fd3a00dea90a4a5ea543652931b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621241
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>