[scrutiny] Fix directionality of routes
The documentation in the Routes collection states that the source of a
route should be the component providing the capability. However, the
collector constructs routes with the source being the component that
uses the capability, effectively reversing the directionality. The
capability source was also confusingly called "target".

This change fixes the name to make it clear it is the source of the
capability, and swaps the source and destination in the Route
assignment. It also introduces two tests to observe the directionality
of routes, which was not previously covered.

Test: fx test scrutiny_plugins_lib_test
Fixed: 90927
Change-Id: I525908c21f750cffd32f421c98422a0b5b2f7343
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626430
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Commit-Queue: Devon O'Dell <dhobsd@google.com>