[brcmfmac] Clear FW connection state.
FW connection state sometimes does not get cleared
when a connection attempt fails. This change
issues C_DOWN on the client IF which is expected
to clear up the FW connection state.

Bug: 90650

Test: Added a SIM test to verify C_DOWN and C_UP are
set in FW during client disconnect. Also manually
tested by bringing down the client connection
and checking to see if it affects the SoftAP IF.

Change-Id: I390cfcb0f2613edf1735db979d75a1d628595866
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612221
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Commit-Queue: Karthik Krishnaswami <karthikrish@google.com>