[fidl_driver] Arenas for driver transport
This CL wires through the driver arena parameter so that arenas can be
used in requests and responses. The API used here is not meant to be
final, but merely to make a functional interface.

A disabled end-to-end test is included in this CL. The test typically
passes on non-ASAN builds, but is flaky and involves a use-after-free
in driver runtime teardown. This use-after-free is not new to this CL
and the CL is being sent for review despite the use-after-free to
prevent code rot and to make it possible for other changes to continue
to build on top of it.

Change-Id: I42099cb4bde79a72a9af05daa2d1d5f03a133d3a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621621
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>