[hub] Structured Config Values are output in the hub
When a component is resolved, its config values become known.
These are sent over the Resolved event to the hub.

The hub has a `resolved/config` folder where it puts all
config fields as a file.

The content of these files is the corresponding config value.

Bug: 86139
Test: fx test component-manager-tests
Test: fx test hub_integration_test
Change-Id: I6f1b73553d8d905a301532b0d9a52176d655b13a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/625103
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>