[driver manager] Refactor driver binding
Move some binding functions out of Coordinator into
a BindDriverManager class.

Remove AttemptBind from the function parameters and
store it in the BindDriverManager. Outside of a test,
the same AttemptBind is being passed around. Expose
an AttemptBind mutator for the test instead.

Rename functions so that the names are less similar
to each other and more explicit on the functionality.

Remove multiple BindDriver() function overloads:
 - Remove BindDriver(Driver*, AttemptBind) since
   this is only used in a test.
 - Rename BindDriver(fbl::RefPtr<Device>&, const char*)
   to BindDriverToDevice() since the functionality is
   completely different.

Move namespace definitions in coordinator.cc to the
top

Bug: 89910
Change-Id: I0b9c8e9c2bb0d7945451eb6a2a7400287eed56ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626209
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Sarah Chan <spqchan@google.com>