[sl4f][modular] Resolve session component before using its protocols
The facade connects to session protocols through a hub directory
that is created when the component is launched. However, the `exec`
subdirectory is not created until the component is *started*, leading
to a race.

Now the facade will ensure the component is resolved, and then connect
to protocols under the `resolved` directory.

Bug: 87456
Change-Id: I488a90a718968a7394f6a4e71455582dd2e08d9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614992
Reviewed-by: Yuan Zhi <yuanzhi@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>
(cherry picked from commit 7016890e9341f92f9dddd9247eec9c28c2170118)