[zxio] Add a zxio_reopen operation with flags
This adds a zxio_reopen wrapper for the io1 Clone or io2 Reopen
operation and defines an option to request a description of the
connection from the server. zxio_clone internally calls zxio_reopen with
the default options.

Currently, reopen is implemented with a fuchsia.io.Node/Clone call. In
the future we'll want to switch this over to the io2 Reopen call.

This also adds tests for both the zxio_clone and the new
zxio_reopen operations with the remote backend.

Bug: 43267
Bug: 77623

Change-Id: I183fd7cb13ab7aad3d6cd866ef4cba1b8b58d0f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627028
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>