[ffx] Additional flags in the start-up template
This CL adds more flags to the start-up template, based on
device_launcher code. Specifically, we add flags for hidpi scaling, gpu,
window size, audio, image files, and serial monitoring, then added a few
predicates for some that were previously hard-coded but shouldn't be.

Based on a dry-run comparison this takes us to parity with the default
setting in fvdl, with these exceptions:
 - grpc port flags
 - fvm image flags
 - networking flags

For the acceleration flag, this checks that KVM is accessible and that the
current user can write to it in order to resolve the "auto" value.

It also updates the sequence of events so that template resolution
happens after file staging and other tasks, but just before the command
is launched. This makes sure the flags contain the most up-to-date
values.

Bug: fxbug.dev/90481
Change-Id: Ic58dd9715820b2ac818d879a41e3b9a4a8acea3f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626985
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>