[terminal] Add 1.25 scale factor bucket
This adds a 1.25 scale factor bucket and moves the
code that is used to determine the scale factor into
the terminal library so it can be used by both virtcon
and the terminal app.

The 1.25 scale factor is useful for 1080p laptop displays
and the workstation product will be updated to take
advantage of it in a follow up change.

The terminal app will be using a hardcoded DPI value of
160 until Flatland API provides what is needed to
dynamically determine DPI.

Change-Id: Ifa082b1a07158fde6c4c3224b206e8c6f9127e25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626662
Reviewed-by: Rob Tsuk <robtsuk@google.com>
Commit-Queue: David Reveman <reveman@google.com>