[fdio] Use zxio to interact with remote nodes in local namespace
The fdio library maintains a local namespace of paths that terminate in
remote objects that communicate with external fuchsia.io servers. This
updates the local namespace logic to use the zxio library for the server
interactions to better encapsulate the FIDL logic. This does increase
the size of a LocalVnode instance to store an arbitrary zxio object
instead of only a channel handle.

This also cleans up the type management in the local connection tracking
a bit to use constructors, destructors, and RefPtr for memory management
instead of raw pointer manipuation.

Change-Id: Ibae9d21d70f5177ad0609675deb3c8211c4ed63f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626801
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>