[ffx] Fix the prebuilt path for AEMU
fxrev.dev/626986 changed the version of AEMU that's being imported to
the Fuchsia SDK prebuilts, which broke the fallback for ffx emu. This CL
updates the fallback code to look for the AEMU binary at the new
location, and also canonicalizes the path to get rid of "../" and such.

It also fixes some indentation in the template, which has no functional
impact.

TESTED=Manually ran against new binary location successfully.

Change-Id: Ib99600bcb2b0499da99b4b840ecf000980422188
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627602
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>