[scrutiny][resolvers] Scrutiny uses PartialAbsoluteMoniker for component comparison
Replace `AbsoluteMoniker` with `PartialAbsoluteMoniker` when Scrutiny
compares an instance moniker with a request moniker while walking
through the component graph. Additionally, instead of translating the
monikers into Strings for comparison, the monikers are compared
directly. Push the `PartialAbsoluteMoniker` of the visited instance in
Scrutiny's moniker list, since it does not include instances already.

This change has a dependency that all components in the component graph
have unique `PartialAbsoluteMoniker` values so that accuracy is not lost
when using it for comparisons. The reason for this change is to remove
AbsoluteMonikers from the public API space.

Bug: 87407
Change-Id: I2c3f066ff8d5062e0826557443b7436555575fad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627000
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Laura Peskin <pesk@google.com>
Reviewed-by: Olivia Nordquist <nolivia@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>