[fidlc] Output transport in JSON IR for client end / server end
In order for fidlgen generators to know if a client end or server end is
associate with drivers or zircon channels, it is necessary to know the
transport. This CL outputs the transport alongside the client end and
server end in the JSON IR in order to accomplish this.

Bug: 56727

Change-Id: I424f5c7142e2f988b06acfd0ac8d0a48736298e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627201
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>