[Recovery] Add new Proxy View navigation controller
ProxyView allows the application to switch views enabling
completely different screens to be shown.

ProxyView pushes new views to a stack and pops them off when
requested by a Carnelian Message.

This is the first of several connected CLs for the on-screen
keyboard which uses ProxyView to show and hide the keyboard.

A change has been made to the carnelian sources because I could
find no way to create a mock ViewAssistantContext.

Bug: b/202857033 b/203107208
Test: New tests added
Change-Id: Ic29df3e35a59b9b2412e7e211f0d4e6daa1a2ecd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620081
Reviewed-by: Mike "mc" Comfoltey <comfoltey@google.com>
Reviewed-by: Rob Tsuk <robtsuk@google.com>
Commit-Queue: Kevin Twidle <kpt@google.com>
Fuchsia-Auto-Submit: Kevin Twidle <kpt@google.com>