[fidl][llcpp] Some helper methods for fidl::Reason
fidl::Reason is an enum with lots of members, some of which have subtle
meanings. Having an enum could be problematic for future extensions
because users might try to exhaustively match on the members.
This CL is an interim step to address the complexity. While one could
still reach for the underlying enum and write `switch case` if they
really need it, it is now discouraged, in comparison to a few new helper
accessors.

I've changed most of the big `switch case`s in-tree to use the accessors
to demonstrate the simplification, but will split user code changes out
when landing.

Bug: 75702

Change-Id: I9edb1b4773c081835adce2a7266f17c6d2359c1d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626807
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>