[ffx] Move environment vars to the args template
This CL adds an "envs" field to the FlagData struct, so that environment
variables can be specified in the start up args template file. We remove
the previously unused env field from the runtime struct, as well as any
references to it. Then we added a few defaults to the template for
DISPLAY and QEMU_AUDIO_DRV, which fix errors or warnings we were seeing
without them, and a template helper to populate fields with environment
value.

To aid in correct DISPLAY behavior, we also output a hint to switch to
--headless if none is set (which usually indicates a text-only session).

Bug: fxbug.dev/90481
Change-Id: I2752b86c2bc192ffcda659c8568dbd7433673f61
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627765
Fuchsia-Auto-Submit: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>