[llcpp] Split out wire messaging declarations
There is a strange ordering of dependencies, where wire_messaging.h is
both the top-level entry point header for messaging and a dependency
of other headers. This makes it harder to avoid dependency cycles with
includes. This CL splits out the forward declarations into a separate
header to reduce some of the complication.

This also fixes a bug, where a call to fidl_write_epitaph may in some
cases be made for non-zircon channel handles.

Change-Id: Id5ff33798afecd9812e10069459eac12baf5ba46
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627464
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>