[llcpp] Support driver client end inside of FIDL types
This CL makes it so driver client_end and server_end, specified in the
FIDL language are generated into fdf::ClientEnd and fdf::ServerEnd.

One of the bigger challenges is dealing with a dependency on the driver
library from the wire type library. To solve this, a flag was added to
the FIDL build rule indicating that the FIDL files use drivers.

Change-Id: Ia806c6acdd0da06ca68f31bfe6c30e1511a1ef59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627462
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>