[Recovery] Add Keys and Keyboard layout code to recovery
The keys have lower, upper case and alt variants.

The keys are brought together in the keyboard which also
contains two text fields, one for a title an one for user input.

This is the second of several connected CLs for the on-screen
keyboard which uses ProxyView to show and hide the keyboard.

Bug: b/202857033 b/203107208
Test: None
Change-Id: I7141f9c5cfa26edc3a0b558fe5714714a2375c05
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620084
Reviewed-by: Rob Tsuk <robtsuk@google.com>
Reviewed-by: Mike "mc" Comfoltey <comfoltey@google.com>
Commit-Queue: Kevin Twidle <kpt@google.com>